requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
pandas
numpy
git+https://github.com/ultralytics/yolov5.git